requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
py4j==0.10.9.7
PySocks==1.7.1
pyspark==3.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
selenium==4.21.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
unicodedata2==15.1.0
Unidecode==1.3.8
urllib3==2.2.1
webdriver-manager==4.0.1
Werkzeug==3.0.3
wsproto==1.2.0
